# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aegis Inspector",
    "description": "\nThe inspection and governance engine behind the **Aegis** Zero-Trust\nResponsible AI Gateway.\n\nThis service decides whether a prompt may be transmitted, screens model\nresponses on the way back, and records the platform's decision history. It is\nconsumed by the Aegis Gateway, never by end clients directly.\n\n### What this service does not claim\n\n* Detection is **heuristic and incomplete**. Prompt-injection defence covers\n  known OWASP LLM01 patterns and is bypassable by obfuscation.\n* Grounded Response Verification is a **support score**, not a guarantee of\n  factual correctness.\n* Energy and CO2 figures are **estimates** from configurable assumptions, not\n  measurements.\n* Audit output is **transaction evidence** supporting a deployer's own\n  record-keeping. It is not a conformity assessment and does not establish\n  legal compliance.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/audit/events": {
      "post": {
        "tags": [
          "audit"
        ],
        "summary": "Record one finalized request",
        "description": "Called by the Gateway after a response completes, including after a stream ends. Idempotent on `request_id`.\n\nA failed audit write must never fail the user's request -- the Gateway should fire-and-forget and increment its own `audit_write_failures` counter.\n\n`user_ref` is hashed on arrival and the raw value is discarded. The model has no field for prompt text, keys or passwords, so they cannot be persisted even by accident.",
        "operationId": "write_audit_event_audit_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/events": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Paginated decision log",
        "operationId": "list_audit_events_api_audit_events_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Tenant Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "policy_action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Policy Action"
            }
          },
          {
            "name": "since_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8760,
              "minimum": 1,
              "default": 24,
              "title": "Since Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit/report": {
      "get": {
        "tags": [
          "audit"
        ],
        "summary": "Structured audit export",
        "description": "Returns structured sections the Gateway renders. Carries a disclaimer stating that this is transaction evidence supporting a deployer's own obligations, not a conformity assessment.",
        "operationId": "audit_report_api_audit_report_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "default",
              "title": "Tenant Id"
            }
          },
          {
            "name": "since_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 8760,
              "minimum": 1,
              "default": 168,
              "title": "Since Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuditEventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEventRecord"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "AuditEventPage"
      },
      "AuditEventRecord": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "policy_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Action"
          },
          "rules_fired": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules Fired"
          },
          "pii_count": {
            "type": "integer",
            "title": "Pii Count",
            "default": 0
          },
          "secret_count": {
            "type": "integer",
            "title": "Secret Count",
            "default": 0
          },
          "injection_detected": {
            "type": "boolean",
            "title": "Injection Detected",
            "default": false
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "failover_used": {
            "type": "boolean",
            "title": "Failover Used",
            "default": false
          },
          "grounding_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grounding Score"
          },
          "latency_total_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Total Ms"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "request_id",
          "tenant_id",
          "timestamp"
        ],
        "title": "AuditEventRecord",
        "description": "Read-side projection. Deliberately narrower than the write model."
      },
      "AuditEventRequest": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id",
            "default": "default"
          },
          "user_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Ref"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "routed_complexity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routed Complexity"
          },
          "latency_total_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Total Ms"
          },
          "latency_inspect_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Inspect Ms"
          },
          "latency_provider_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Provider Ms"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          },
          "cache_similarity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cache Similarity"
          },
          "pii_detected": {
            "type": "boolean",
            "title": "Pii Detected",
            "default": false
          },
          "pii_count": {
            "type": "integer",
            "title": "Pii Count",
            "default": 0
          },
          "pii_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pii Types"
          },
          "secret_detected": {
            "type": "boolean",
            "title": "Secret Detected",
            "default": false
          },
          "secret_count": {
            "type": "integer",
            "title": "Secret Count",
            "default": 0
          },
          "secret_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secret Types"
          },
          "injection_detected": {
            "type": "boolean",
            "title": "Injection Detected",
            "default": false
          },
          "injection_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Injection Score"
          },
          "policy_action": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Decision"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Version"
          },
          "rules_fired": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rules Fired"
          },
          "failover_used": {
            "type": "boolean",
            "title": "Failover Used",
            "default": false
          },
          "failover_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failover From"
          },
          "failover_to": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failover To"
          },
          "egress_flagged": {
            "type": "boolean",
            "title": "Egress Flagged",
            "default": false
          },
          "egress_categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Egress Categories"
          },
          "grounding_enabled": {
            "type": "boolean",
            "title": "Grounding Enabled",
            "default": false
          },
          "grounding_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grounding Score"
          },
          "grounding_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundingStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "estimated_cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Cost Usd"
          },
          "estimated_savings_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Savings Usd"
          },
          "estimated_energy_wh": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Energy Wh"
          },
          "estimated_co2_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Co2 G"
          },
          "review_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "AuditEventRequest"
      },
      "AuditEventResponse": {
        "properties": {
          "stored": {
            "type": "boolean",
            "title": "Stored"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stored",
          "request_id"
        ],
        "title": "AuditEventResponse"
      },
      "AuditReport": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "window_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Start"
          },
          "window_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window End"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/AuditReportSection"
            },
            "type": "array",
            "title": "Sections"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "generated_at",
          "tenant_id",
          "sections",
          "disclaimer"
        ],
        "title": "AuditReport"
      },
      "AuditReportSection": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requirement"
          },
          "rows": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Rows"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title"
        ],
        "title": "AuditReportSection"
      },
...
      "Decision": {
        "type": "string",
        "enum": [
          "ALLOW",
          "SANITIZE",
          "WARN",
          "BLOCK"
        ],
        "title": "Decision"
      },
...
      "GroundingStatus": {
        "type": "string",
        "enum": [
          "GROUNDED",
          "REVIEW",
          "UNGROUNDED",
          "SKIPPED"
        ],
        "title": "GroundingStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}